- fixed some bugs - some code cleanup - removed window for adding financial operations

+++ SyZaFi/DBConnection.cs
@@ -609,7 +609,7 @@ namespace SyZaFi
             }
         }
 
-        public void MoveEmployeeToRemovedEmployees(string firstName, string lastName, string fathersName, string mothersName, string birthday, string placeOfResidence, string correspondenceAddress, string education, string historyOfEmployment, int pesel, string childrensNames, string childrensBirthdays)
+        public void MoveEmployeeToRemovedEmployees(string firstName, string lastName, string fathersName, string mothersName, string birthday, string placeOfResidence, string correspondenceAddress, string education, string historyOfEmployment, long pesel, string childrensNames, string childrensBirthdays)
         {
             string query = "INSERT INTO deletedemployees (firstName, lastName, fathersName, mothersName, birthday, placeOfResidence, correspondenceAddress, education, historyOfEmployment, pesel, childrensNames, childrensBirthdays)" +
                 " VALUES('" + firstName + "', '" + lastName + "', '" + fathersName + "', '" + mothersName + "', '" + birthday + "', '" + placeOfResidence + "', '" + correspondenceAddress + "', '" + education + "', '" + historyOfEmployment + "', '" + pesel + "', '" + childrensNames + "', '" + childrensBirthdays + "')";
